+++ src/main/java/store/model/store/promotion/Promotion.java
@@ -1,4 +1,4 @@
-package store.model;
+package store.model.store.promotion;
 
 import java.time.LocalDateTime;
