+++ src/main/java/store/util/reader/StoreFileReader.java
@@ -1,4 +1,4 @@
-package store.util;
+package store.util.reader;
 
 import java.io.BufferedReader;
 import java.io.IOException;
