+++ src/test/java/store/model/order/PurchaseOrderTest.java
@@ -1,4 +1,4 @@
-package store.model;
+package store.model.order;
 
 import static camp.nextstep.edu.missionutils.DateTimes.now;
 
@@ -7,6 +7,10 @@ import java.util.Map;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import store.model.store.StoreRoom;
+import store.model.store.product.Product;
+import store.model.store.product.Products;
+import store.model.store.promotion.Promotion;
 
 @DisplayName("사용자 구매 목록 모델 테스트")
 class PurchaseOrderTest {
